Move functions to create DOM elements to helpers
