+++ src/scripts/modules/views/project.js
@@ -2,15 +2,11 @@
 import projects from '../../index';
 // eslint-disable-next-line import/no-cycle
 import todoController from '../controller/todo';
+// eslint-disable-next-line import/no-cycle
 import renderProjects from './renderProjects';
+import { myCreateElement, generateBtn, generateFormField } from './helpers';
 
 const projectView = () => {
-  const myCreateElement = (el, className) => {
-    const element = document.createElement(el);
-    element.classList.add(className);
-    return element;
-  };
-
   const openCurrentTabAndContainer = (project) => {
     document.querySelectorAll('.project-tabs__item').forEach(tab => {
       if (project.id === tab.getAttribute('data-id')) {
@@ -22,28 +18,6 @@ const projectView = () => {
         container.classList.add('active');
       }
     });
-  };
-
-  const generateFormField = (labelName, type, inputName) => {
-    const field = myCreateElement('div', 'form__field');
-    const label = myCreateElement('label');
-    label.textContent = labelName;
-
-    const input = document.createElement('input');
-    input.setAttribute('type', type);
-    input.setAttribute('name', inputName);
-    input.required = true;
-
-    field.appendChild(label);
-    field.appendChild(input);
-
-    return [field, input];
-  };
-
-  const generateBtn = (text) => {
-    const button = myCreateElement('button', 'btn');
-    button.textContent = text;
-    return button;
   };
 
   const generateForm = (todo, formClass) => {
